Reverse Infinite Scrolling Milestone
Okay, so "reverse" infinite scrolling (dynamically loading PAST messages) is incredibly difficult.  This is because some browsers (looking at you, Safari) don't lock the scroll position when adding new content to the window above.

So this is a (short term?) solution that requires a click to manually load older posts.  Gonna try this UX for a while and see how it feels.