Audit user permissions
Now that we are (potentially) forwarding users and streams that have moved to new servers, we need to verify every access point to ensure that we're forwarding them correctly.

This is done in wrappers like `WithUser` `WithStream` and similar -- so every handler that touches Users or Streams MUST be called though this middleware.